Fix and refactor user icons
