+++ lib/gollum/public/gollum/javascript/app.js
@@ -1,4 +1,5 @@
 //= require jquery-1.7.2.min
+//= require identicon_canvas
 //= require mousetrap.min
 //= require gollum
 //= require gollum.dialog
