+++ lib/gollum/public/gollum/stylesheets/tables.scss
@@ -2,6 +2,11 @@
 
 /* @section history */
 
+#user-icons {
+  a, img, span {
+    vertical-align: middle;
+  }
+}
 
 .history #footer {
   margin-bottom: 7em;
