+++ lib/gollum/templates/history.mustache
@@ -15,7 +15,7 @@
 	    	{{#versions}}
 				<li class="Box-row border-top Box-row--hover-gray d-flex flex-items-center">
 					<span class="pr-2"><input class="checkbox" type="checkbox" name="versions[]" value="{{id}}"></span>
-					<span class="float-left col-2">{{>author_template}}</span>
+					<span class="float-left col-2" id="user-icons">{{>author_template}}</span>
 					<span class="flex-auto col-1 text-gray-light">{{date}}</span>
 					<span class="flex-auto col-5">{{message}}</span>
 					<span class="pl-4 float-right">[<a href="{{base_url}}/{{escaped_url_path}}/{{id}}" title="View commit">{{id7}}</a>]</span>
