+++ lib/gollum/templates/history_authors/gravatar.mustache
@@ -1,5 +1,3 @@
-<a href="javascript:void(0)">
 <img src="https://secure.gravatar.com/avatar/{{gravatar}}?s=16"
     alt="avatar: {{author}}" class="mini-gravatar"/>
-    {{author}}
-</a>
+&nbsp;<a href="https://gravatar.com/{{gravatar}}">{{author}}</a>
