+++ lib/gollum/templates/history_authors/identicon.mustache
@@ -1,5 +1,2 @@
-<a href="javascript:void(0)">
-  <img src="{{#sprockets_image_path}}man_24.png{{/sprockets_image_path}}" alt="avatar: {{author}}"
-    class="mini-gravatar identicon" data-identicon="{{identicon}}"/>
-  {{author}}
-</a>
+<img src="" alt="avatar: {{author}}" class="identicon" data-identicon="{{identicon}}"/>
+&nbsp;<span>{{author}}</span>
