+++ lib/gollum/templates/history_authors/none.mustache
@@ -1,1 +1,1 @@
-    {{author}}
+{{author}}
