+++ lib/gollum/templates/latest_changes.mustache
@@ -11,7 +11,7 @@
 <div class="Box flex-auto">
 	{{#versions}}
 		<div class="Box-row Box-row--hover-gray border-top d-flex flex-items-center">
-			<span class="float-left col-2">{{>author_template}}</span>
+			<span class="float-left col-2" id="user-icons">{{>author_template}}</span>
 			<span class="flex-auto col-1 text-gray-light">{{date}}</span>
 			<span class="flex-auto col-7">{{message}}<br/>
 			{{#files}}
