+++ lib/gollum/templates/layout.mustache
@@ -26,9 +26,6 @@
 
   </script>
   {{#sprockets_javascript_tag}}app{{/sprockets_javascript_tag}}
-  {{#use_identicon}}
-  {{#sprockets_javascript_tag}}identicon_canvas{{/sprockets_javascript_tag}}
-  {{/use_identicon}}
   {{#mathjax}}
   {{^mathjax_config}}
   <script type="text/javascript">
